+++ src/pages/Login.tsx
@@ -1,48 +1,18 @@
 import { EyeIcon, EyeSlashIcon } from '@heroicons/react/24/outline';
-import { requestLogin } from '../apis/auth';
-import type { LoginRequest } from '../types/auth';
-import { useState } from 'react';
-import { useNavigate, Link } from 'react-router';
-
+import useLogin from '../hooks/useLogin';
+import { Link } from 'react-router';
 
 function Login() {
-  const navigate = useNavigate();
-
-  const [errorMsg, setErrorMsg] = useState<string | null>(null);
-
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [showPw, setShowPw] = useState(false);
-
-  const togglePw = () => setShowPw((prev) => !prev);
-
-  const login = async (loginData: LoginRequest): Promise<void> => {
-    try {
-      const data = await requestLogin(loginData);
-      // if 제거할 시 컴파일러에서 빨간 줄 띄워집니다.
-      if (data.success) {
-        localStorage.setItem("accessToken", data.data.accessToken);
-        localStorage.setItem("user", JSON.stringify(data.data.user));
-        navigate("/");
-      }
-    } catch (error: any) {
-      setErrorMsg("아이디 또는 비밀번호가 올바르지 않습니다.");
-    }
-  }
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (!username || !password) {
-      alert("아이디와 비밀번호를 확인하세요.")
-      return;
-    }
-
-    const loginData: LoginRequest = { username, password };
-
-    setErrorMsg(null);
-    login(loginData);
-  }
+  const {
+    username,
+    password,
+    showPw,
+    errorMsg,
+    setUsername,
+    setPassword,
+    togglePw,
+    handleSubmit
+  } = useLogin();
 
   return (
     <div className="max-w-[1190px] mx-auto px-2">
